move some parse tests into own files
